requirements.txt
Flask
gunicorn
customtkinter
Pillow
pypdf
google-genai
python-dotenv
